createpageupdates

+++ shop/models.py
@@ -5,7 +5,7 @@ from django.db import models
 class productmodel(models.Model):
     Title= models.CharField(max_length=225,name='TITLE',null=False)
     Description=models.TextField(max_length=500,name='DESCRIPTION',null=True)
-    Price=models.IntegerField(name="PRICE",null=False)
+    Price=models.DecimalField(name="PRICE",null=False,decimal_places=4,max_digits=25)
     Rating=models.IntegerField(name='RATINGS',default=0)
     imagepath=models.ImageField(name="IMAGE",upload_to='product_images/', null=True)
     Product_Category=models.CharField(name="PRODUCT_CATEGORY",max_length=500, null=True)
